Unify SyntaxTreeCreator
