+++ java-squid/src/main/java/org/sonar/java/parser/sslr/ActionParser.java
@@ -43,6 +43,7 @@ import org.sonar.sslr.parser.ParseRunner;
 import org.sonar.sslr.parser.ParsingResult;
 
 import javax.annotation.Nullable;
+
 import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
@@ -90,7 +91,7 @@ public class ActionParser {
       }
     }
 
-    this.syntaxTreeCreator = new SyntaxTreeCreator<>(treeFactory, grammarBuilderInterceptor);
+    this.syntaxTreeCreator = new SyntaxTreeCreator<>(treeFactory, grammarBuilderInterceptor, new JavaNodeBuilder());
 
     b.setRootRule(rootRule);
     this.rootRule = rootRule;
