+++ java-squid/src/main/java/org/sonar/java/parser/sslr/SyntaxTreeCreator.java
@@ -24,16 +24,11 @@ import com.google.common.base.Throwables;
 import com.google.common.collect.Lists;
 import com.sonar.sslr.api.AstNode;
 import com.sonar.sslr.api.GenericTokenType;
-import com.sonar.sslr.api.Rule;
 import com.sonar.sslr.api.Token;
 import com.sonar.sslr.api.TokenType;
 import com.sonar.sslr.api.Trivia;
 import com.sonar.sslr.api.Trivia.TriviaKind;
-import org.sonar.java.model.InternalSyntaxSpacing;
-import org.sonar.java.model.InternalSyntaxToken;
-import org.sonar.java.model.InternalSyntaxTrivia;
 import org.sonar.java.parser.sslr.ActionParser.GrammarBuilderInterceptor;
-import org.sonar.plugins.java.api.tree.SyntaxTrivia;
 import org.sonar.sslr.grammar.GrammarRuleKey;
 import org.sonar.sslr.internal.grammar.MutableParsingRule;
 import org.sonar.sslr.internal.matchers.ParseNode;
@@ -49,21 +44,27 @@ public class SyntaxTreeCreator<T> {
 
   private final Object treeFactory;
   private final GrammarBuilderInterceptor mapping;
+  private final NodeBuilder nodeBuilder;
 
   private final Token.Builder tokenBuilder = Token.builder();
   private final List<Trivia> trivias = Lists.newArrayList();
 
   private Input input;
 
-  public SyntaxTreeCreator(Object treeFactory, GrammarBuilderInterceptor mapping) {
+  public SyntaxTreeCreator(Object treeFactory, GrammarBuilderInterceptor mapping, NodeBuilder nodeBuilder) {
     this.treeFactory = treeFactory;
     this.mapping = mapping;
+    this.nodeBuilder = nodeBuilder;
   }
 
   public T create(ParseNode node, Input input) {
     this.input = input;
     this.trivias.clear();
-    return (T) visit(node);
+    T result = (T) visit(node);
+    if (result instanceof AstNode) {
+      ((AstNode) result).hasToBeSkippedFromAst();
+    }
+    return result;
   }
 
   private Object visit(ParseNode node) {
@@ -101,6 +102,7 @@ public class SyntaxTreeCreator<T> {
     List<Object> convertedChildren = Lists.newArrayList();
     for (ParseNode child : children) {
       Object result = visit(child);
+
       if (result != null) {
         // FIXME to remove aafter full migration: Allow to skip optional nodes that are supposed to bw skipped from the AST
         if ((result instanceof Optional && ((Optional) result).isPresent() && ((Optional) result).get() instanceof AstNode) && ((AstNode) ((Optional) result).get()).hasToBeSkippedFromAst()){
@@ -128,7 +130,10 @@ public class SyntaxTreeCreator<T> {
 
     Method method = mapping.actionForRuleKey(ruleKey);
     if (method == null) {
-      return createNonTerminal(ruleKey, rule, convertedChildren, node.getStartIndex(), node.getEndIndex());
+      int startIndex = node.getStartIndex();
+      int endIndex = node.getEndIndex();
+
+      return nodeBuilder.createNonTerminal(rule.getRuleKey(), rule, convertedChildren, startIndex, endIndex);
     }
 
     try {
@@ -142,18 +147,7 @@ public class SyntaxTreeCreator<T> {
     }
   }
 
-  private static Object createNonTerminal(GrammarRuleKey ruleKey, Rule rule, List<Object> children, int startIndex, int endIndex) {
-    for (Object child : children) {
-      if (child instanceof InternalSyntaxToken) {
-        InternalSyntaxToken syntaxToken = (InternalSyntaxToken) child;
-        syntaxToken.setGrammarRuleKey(ruleKey);
-        return child;
-      }
-    }
-    return new InternalSyntaxSpacing(startIndex, endIndex);
-  }
-
-  private InternalSyntaxToken visitTerminal(ParseNode node) {
+  private Object visitTerminal(ParseNode node) {
     TokenType type = null;
     if (node.getMatcher() instanceof TriviaExpression) {
       TriviaExpression ruleMatcher = (TriviaExpression) node.getMatcher();
@@ -179,25 +173,8 @@ public class SyntaxTreeCreator<T> {
         return null;
       }
     }
-    InternalSyntaxToken internalSyntaxToken =
-      createTerminal(input, node.getStartIndex(), node.getEndIndex(), Collections.unmodifiableList(trivias), type);
+    Object result = nodeBuilder.createTerminal(input, node.getStartIndex(), node.getEndIndex(), Collections.unmodifiableList(trivias), type);
     trivias.clear();
-    return internalSyntaxToken;
-  }
-
-  private static InternalSyntaxToken createTerminal(Input input, int startIndex, int endIndex, List<Trivia> trivias, TokenType type) {
-    boolean isEof = GenericTokenType.EOF == type;
-    LineColumnValue lineColumnValue = tokenPosition(input, startIndex, endIndex);
-    return new InternalSyntaxToken(lineColumnValue.line, lineColumnValue.column, lineColumnValue.value,
-      createTrivias(trivias), startIndex, endIndex, isEof);
-  }
-
-  private static List<SyntaxTrivia> createTrivias(List<Trivia> trivias) {
-    List<SyntaxTrivia> result = Lists.newArrayList();
-    for (Trivia trivia : trivias) {
-      Token trivialToken = trivia.getToken();
-      result.add(InternalSyntaxTrivia.create(trivialToken.getValue(), trivialToken.getLine(), trivialToken.getColumn()));
-    }
     return result;
   }
 
@@ -209,24 +186,6 @@ public class SyntaxTreeCreator<T> {
     tokenBuilder.setURI(input.uri());
     String value = input.substring(node.getStartIndex(), node.getEndIndex());
     tokenBuilder.setValueAndOriginalValue(value);
-  }
-
-  private static LineColumnValue tokenPosition(Input input, int startIndex, int endIndex) {
-    int[] lineAndColumn = input.lineAndColumnAt(startIndex);
-    String value = input.substring(startIndex, endIndex);
-    return new LineColumnValue(lineAndColumn[0], lineAndColumn[1] -1, value);
-  }
-
-  private static class LineColumnValue {
-    final int line;
-    final int column;
-    final String value;
-
-    private LineColumnValue(int line, int column, String value) {
-      this.line = line;
-      this.column = column;
-      this.value = value;
-    }
   }
 
 }
